Feat : review 반환시 json key Name 변경

+++ src/main/java/com/enjoytrip/attraction/dao/AttractionReviewMapper.java
@@ -14,7 +14,7 @@ public interface AttractionReviewMapper {
     List<AttractionReview> getReviewsByUserId(String userId);
 
     AttractionReview getOneReview(String reviewId);
-    int writeReview(AttractionReview review);
+    Long writeReview(AttractionReview review);
     int deleteReview(String reviewId);
 
     AttractionReviewScoreDto getEvaluation(String attractionId);
